reorganised repo to be more organised #31
