+++ README.rst
@@ -34,7 +34,7 @@ required TheSkyX driver see sections below. Assuming everything is set up
 correctly, this is how you would set up the dome control system.
 
 * Navigate to the directory containing the gRPC python server file
-* ``$ cd huntsman-dome/domehunter/protos/src/``
+* ``$ cd huntsman-dome/domehunter/gRPC-server/``
 * Use the -h flag to get information on the available command line flags
 * ``$ python hx2dome.proto_server.py -h ``
 * To run server in simulated hardware mode run with the -sh flag
@@ -112,7 +112,7 @@ Compiling TheSkyX Driver
 ------------------------
 
 The files for compilation and installation are found in the
-``domehunter/protos/`` directory. The relevant files are,
+``domehunter/gRPC-TheSkyX-driver/`` directory. The relevant files are,
 
 
 * ``domelistHuntsmanDome.txt``
@@ -120,10 +120,7 @@ The files for compilation and installation are found in the
 * ``Makefile_[MAC,LINUX]``
 
 The first two are files are used to install the compiled c++
-driver into TheSkyX application directory. You should be
-able to simply run the shell script once the driver is compiled
-and located in the ``domehunter/protos/`` directory, with
-filename ``libHuntsmanDome.so``.
+driver into TheSkyX application directory.
 
 |
 
@@ -132,9 +129,10 @@ In order to compile the driver simply run the makefile recipe for your OS (LINUX
   cd domehunter/protos/
   make -f Makefile_LINUX
 
-This will produce a .so file in the protos directory for Linux and a .dylib file for Mac.
-This file, as well as the ``domelistHuntsmanDome.txt`` file need to be copied into TheSkyX
-application directory. This can be done by running the installation script::
+This will produce a .so file in the ``domehunter/gRPC-TheSkyX-driver/``
+directory for Linux and a .dylib file for Mac. This file, as well as the
+``domelistHuntsmanDome.txt`` file need to be copied into TheSkyX application
+directory. This can be done by running the installation script::
 
   . TheSkyX_LINUX_plugin_install.sh
 
@@ -144,27 +142,33 @@ Replace `LINUX` with `MAC` if installing on a MacOS system and vice versa.
 
 Once the driver is installed in TheSkyX, it can be selected from
 the dome selection menu. Before issuing any commands, start the
-``domehunter/protos/src/hx2dome.proto_server.py`` in a new terminal.
+``domehunter/gRPC-server/hx2dome.proto_server.py`` in a new terminal.
 When you issue a command through TheSkyX, the c++ driver will send
 a remote procedure call through to the gRPC python server.
 
 The gRPC python server can be run in a communication test mode that doesn't
 require any hardware. It will simply return dummy messages back to the TheSkyX
-driver.
+driver. To get help with running the server (in normal or testing mode etc),
+run the following command in terminal,
+
+  cd domehunter/gRPC-server/
+  python hx2dome.proto_server.py -h
 
 gRPC automatically generated files
 ----------------------------------
 
-In the ``domehunter/protos/`` directory there are a number of shell
-scripts. These can be used to generate the gRPC files within the ``src/``
-directory. These scripts contain path variables that may need to be
-adjusted to your local machine. You shouldn't need to worry about
-this as the generated files are committed to the repositry and
+In the ``domehunter/gRPC-TheSkyX-driver/`` directory there are a number
+of shell scripts. These can be used to generate the gRPC files within
+the ``src/`` directory. These scripts contain path variables that may
+need to be adjusted to your local machine. You shouldn't need to worry
+about  this as the generated files are committed to the repository and
 shouldn't need to be generated (I think...?).
 
 The code for the Huntsman dome driver is contained in the
-``domehunter/protos/src`` directory. This directory contains both
-human written files and files automatically generated by gRPC
+``domehunter/gRPC-TheSkyX-driver/src`` directory. The code for the gRPC
+server that runs on the raspberryPi is contained in the
+``domehunter/gRPC-server/`` directory. These directories contain
+both human written files and files automatically generated by gRPC
 tools. The human written files are,
 
 * ``main.cpp`` - establishes main library to X2 driver (mostly copy/paste from example)
@@ -176,8 +180,11 @@ tools. The human written files are,
 
 The remaining cpp and python files are automatically produced
 by gRPC and shouldn't need to be looked at. If for some reason
-you want to generate these files yourself, see the
-*gRPC automatically generated files* section below.
+you want to generate these files yourself, you can use the following shell
+scripts,
+
+* ``domehunter/gRPC-TheSkyX-driver/generate_grpc_cpp_code.sh``
+* ``domehunter/gRPC-server/generate_grpc_python_code.sh``
 
 
 Python RaspberryPi Component
