+++ domehunter/gRPC-TheSkyX-driver/hx2dome.proto
