+++ domehunter/gRPC-TheSkyX-driver/src/hx2dome.grpc.pb.cpp
