+++ domehunter/gRPC-TheSkyX-driver/src/hx2dome.pb.cpp
