+++ domehunter/gRPC-TheSkyX-driver/src/licensedinterfaces/basicstringinterface.h
