+++ domehunter/gRPC-TheSkyX-driver/src/licensedinterfaces/deviceinfointerface.h
