+++ domehunter/gRPC-TheSkyX-driver/src/licensedinterfaces/linkinterface.h
