+++ domehunter/gRPC-TheSkyX-driver/src/licensedinterfaces/loggerinterface.h
