+++ domehunter/gRPC-TheSkyX-driver/src/licensedinterfaces/sberrorx.h
