+++ domehunter/gRPC-TheSkyX-driver/src/licensedinterfaces/tickcountinterface.h
