+++ domehunter/gRPC-TheSkyX-driver/src/x2dome.h
