+++ domehunter/gRPC-server/generate_grpc_python_code.sh
@@ -5,8 +5,8 @@ if [ "$1" == "clean" ]; then
 	rm src/*pb2.py
 else
 	HDOME_PATH="$HOME/Documents/REPOS"
-	PROTOS_PATH="$HDOME_PATH/huntsman-dome/domehunter/protos/src/"
-	PROTO_FILE="$HDOME_PATH/huntsman-dome/domehunter/protos/src/hx2dome.proto"
+	PROTOS_PATH="$HDOME_PATH/huntsman-dome/domehunter/gRPC-TheSkyX-driver/"
+	PROTO_FILE="$HDOME_PATH/huntsman-dome/domehunter/gRPC-TheSkyX-driver/hx2dome.proto"
 
 	echo -e "\nGenerating GRPC Python code\n"
 
@@ -14,8 +14,8 @@ else
 
 	python -m grpc_tools.protoc -I=$PROTOS_PATH --python_out=. --grpc_python_out=. $PROTO_FILE
 
-	echo -e "Moving generated GRPC Python code to src/\n"
-	mv *pb2* src/
+	#echo -e "Moving generated GRPC Python code to src/\n"
+	#mv *pb2* src/
 
 	echo -e "Done.\n"
 fi
